+++ test/tracing/httpServer/test.js
@@ -18,7 +18,8 @@ describe('tracing/httpServer', function() {
   this.timeout(config.getTestTimeout());
 
   agentControls.registerTestHooks({
-    extraHeaders: ['UsEr-AgEnt']
+    extraHeaders: ['UsEr-AgEnt'],
+    secretsList: ['secret', 'Enigma', 'CIPHER']
   });
 
   var controls = new Controls({
@@ -40,7 +41,26 @@ describe('tracing/httpServer', function() {
           return agentControls.getSpans()
           .then(function(spans) {
             utils.expectOneMatching(spans, function(span) {
+              expect(span.n).to.equal('node.http.server');
               expect(span.data.http.header['user-agent']).to.equal(userAgent);
+            });
+          });
+        });
+      });
+  });
+
+  it('must remove secrets from query parameters', function() {
+    return controls.sendRequest({
+      method: 'GET',
+      path: '/?param1=value1&TheSecreT=classified&param2=value2&enIgmAtic=occult&param3=value4&cipher=veiled',
+    })
+      .then(function() {
+        return utils.retry(function() {
+          return agentControls.getSpans()
+          .then(function(spans) {
+            utils.expectOneMatching(spans, function(span) {
+              expect(span.n).to.equal('node.http.server');
+              expect(span.data.http.params).to.equal('param1=value1&param2=value2&param3=value4');
             });
           });
         });